hashi_monitor

Module audit

Auditor implementations. Auditor implementations. Goal: Attempt to match all withdrawals with a guardian event inside the input time window, even if some corresponding events for that withdrawal occur outside the window. Core workflow: - User inputs a window (either just start or both start & end). - Guardian window is authoritative; Sui uses relaxed bounds for predecessor checks. - At a desired frequency, auditors do: - advance cursors - call if wsm.is_in_audit_window() { wsm.violations(&cursors) } to identify errors. - Currently we also report orphan E1 findings when they fall in the user window.

Note that we do not use audit window gating for deposits because there is no risk of false violation flagging.

Re-exports

pub use batch::BatchAuditor;

pub use continuous::ContinuousAuditor;

Modules

batch

continuous

Structs

AuditorCore

ProgressWatermarks

Traits

AuditWindow

Functions

log_findings