hashi_monitor::rpc

Module guardian

Structs

GuardianPollerCore

Reusable S3 poller core with attestation and signature checks. Meant to be used for either withdrawal or heartbeat logs. Idea: Since guardian can write out of order and S3 ListObjectVersions only supports lexicographic cursors, we read from an S3 directory only after we are certain that all writes to it finish. E.g., 12-1 PM bucket is read at 1 PM + DIR_WRITES_COMPLETION_DELAY, e.g., 1:10 PM.

GuardianWithdrawalsPoller

Enums

GuardianLogDir